Use chokidar option awaitWriteFinish
If there is a new file, the dev CLI will wait for the file write to end, and also wait some more time.

This improves Explosiv's relations with other dev build steps. Like before, using Stylus to build CSS files was messy, because the CLI started building before Stylus is done doing its' thing which led to not opening a file because Stylus was building to it at the same moment.